(un)install only when present/absent

+++ libexec/steps/docker.sh
@@ -19,42 +19,44 @@ docker() {
             done
             ;;
         "install")
-            lsb_dist=$(primer_distribution)
-            case "$lsb_dist" in
-                alpine)
-                    primer_packages add docker
-                    ;;
-                clear*linux*)
-                    primer_packages add containers-basic
-                    ;;
-                *)
-                    # Ensure we have curl
-                    primer_dependency curl
+            if ! [ -x "$(command -v dockerd)" ]; then
+                lsb_dist=$(primer_distribution)
+                case "$lsb_dist" in
+                    alpine)
+                        primer_packages add docker
+                        ;;
+                    clear*linux*)
+                        primer_packages add containers-basic
+                        ;;
+                    *)
+                        # Ensure we have curl
+                        primer_dependency curl
 
-                    # Download the installation script from getdocker
-                    yush_info "Downloading and running Docker installation script from $(yush_yellow "$DOCKER_GET_URL"), this is a security risk"
-                    _get=$(mktemp)
-                    curl -fsSL "$DOCKER_GET_URL" -o "$_get"
-                    sh "$_get"
+                        # Download the installation script from getdocker
+                        yush_info "Downloading and running Docker installation script from $(yush_yellow "$DOCKER_GET_URL"), this is a security risk"
+                        _get=$(mktemp)
+                        curl -fsSL "$DOCKER_GET_URL" -o "$_get"
+                        sh "$_get"
 
-                    # On debian we try to ensure that we have the proper
-                    # repository so we have some degree of security.
-                    case "$lsb_dist" in
-                        ubuntu|*bian)
-                            yush_info "Verifying Docker GPG short key against: $(yush_green $DOCKER_APT_GPG)"
-                            dkey=$(apt-key list | grep -e "Docker" -e "docker\.com" -B 1 | head -1 | awk '{print $9$10}')
-                            if [ "$dkey" != "$DOCKER_APT_GPG" ]; then
-                                abort "System might have been compromised, installed short GPG key for Docker was: $dkey"
-                            fi
-                            ;;
-                        *)
-                            yush_warn "Cannot verify proper package provider on $lsb_dist"
-                            ;;
-                    esac
+                        # On debian we try to ensure that we have the proper
+                        # repository so we have some degree of security.
+                        case "$lsb_dist" in
+                            ubuntu|*bian)
+                                yush_info "Verifying Docker GPG short key against: $(yush_green $DOCKER_APT_GPG)"
+                                dkey=$(apt-key list | grep -e "Docker" -e "docker\.com" -B 1 | head -1 | awk '{print $9$10}')
+                                if [ "$dkey" != "$DOCKER_APT_GPG" ]; then
+                                    abort "System might have been compromised, installed short GPG key for Docker was: $dkey"
+                                fi
+                                ;;
+                            *)
+                                yush_warn "Cannot verify proper package provider on $lsb_dist"
+                                ;;
+                        esac
 
-                    rm -f "$_get"
-                    ;;
-            esac
+                        rm -f "$_get"
+                        ;;
+                esac
+            fi
 
             # Start docker and make sure it will always start
             if [ -x "$(command -v dockerd)" ]; then
@@ -75,42 +77,42 @@ docker() {
                 fi
                 yush_info "Disabling docker daemon at start"
                 primer_service disable docker
+
+                lsb_dist=$(primer_distribution)
+                case "$lsb_dist" in
+                    alpine)
+                        primer_packages del docker
+                        ;;
+                    clear*linux*)
+                        primer_packages del containers-basic
+                        ;;
+                    ubuntu|*bian)
+                        primer_packages del docker-ce-cli docker-ce
+                        dkey_present=$(apt-key list | grep -e "Docker" -e "docker\.com" -B 1)
+                        if [ -n "$dkey_present" ]; then
+                            yush_info "Removing docker GPG key"
+                            dkey=$(echo "$dkey_present" | head -1 | awk '{print $9$10}')
+                            $PRIMER_SUDO apt-key del $dkey
+                        fi
+
+                        if [ -f "/etc/apt/sources.list.d/docker.list" ]; then
+                            yush_info "Removing repo list /etc/apt/sources.list.d/docker.list"
+                            $PRIMER_SUDO rm -f /etc/apt/sources.list.d/docker.list
+                        fi
+
+                        if grep -q docker /etc/apt/sources.list; then
+                            yush_info "Removing docker from main repo list at /etc/apt/sources.list"
+                            listtemp=$(mktemp)
+                            grep -v "docker" /etc/apt/sources.list > "$listtemp"
+                            primer_root_ownership "$listtemp" /etc/apt/sources.list
+                            $PRIMER_SUDO mv "$listtemp" /etc/apt/sources.list
+                        fi
+                        ;;
+                    *)
+                        yush_warn "Cannot remove docker on $lsb_dist"
+                        ;;
+                esac
             fi
-
-            lsb_dist=$(primer_distribution)
-            case "$lsb_dist" in
-                alpine)
-                    primer_packages del docker
-                    ;;
-                clear*linux*)
-                    primer_packages del containers-basic
-                    ;;
-                ubuntu|*bian)
-                    primer_packages del docker-ce-cli docker-ce
-                    dkey_present=$(apt-key list | grep -e "Docker" -e "docker\.com" -B 1)
-                    if [ -n "$dkey_present" ]; then
-                        yush_info "Removing docker GPG key"
-                        dkey=$(echo "$dkey_present" | head -1 | awk '{print $9$10}')
-                        $PRIMER_SUDO apt-key del $dkey
-                    fi
-
-                    if [ -f "/etc/apt/sources.list.d/docker.list" ]; then
-                        yush_info "Removing repo list /etc/apt/sources.list.d/docker.list"
-                        $PRIMER_SUDO rm -f /etc/apt/sources.list.d/docker.list
-                    fi
-
-                    if grep -q docker /etc/apt/sources.list; then
-                        yush_info "Removing docker from main repo list at /etc/apt/sources.list"
-                        listtemp=$(mktemp)
-                        grep -v "docker" /etc/apt/sources.list > "$listtemp"
-                        primer_root_ownership "$listtemp" /etc/apt/sources.list
-                        $PRIMER_SUDO mv "$listtemp" /etc/apt/sources.list
-                    fi
-                    ;;
-                *)
-                    yush_warn "Cannot remove docker on $lsb_dist"
-                    ;;
-            esac
             ;;
     esac
 }
